+++ test/runtests.jl
@@ -10,19 +10,21 @@ sets_to_test = [(~ = isapprox, f = sin ∘ *, op = +,
                     itrs = ([1 => "a", 2 => "b", 3 => "c", 4 => "d", 5 => "e"],),
                     init = "")]
 
-ChunkedGreedy(;kwargs...) = GreedyScheduler(;kwargs...)
+ChunkedGreedy(; kwargs...) = GreedyScheduler(; kwargs...)
 
 @testset "Basics" begin
     for (; ~, f, op, itrs, init) in sets_to_test
         @testset "f=$f, op=$op, itrs::$(typeof(itrs))" begin
             @testset for sched in (
-                StaticScheduler, DynamicScheduler, GreedyScheduler, DynamicScheduler{OhMyThreads.Schedulers.NoChunking}, SerialScheduler, ChunkedGreedy)
+                StaticScheduler, DynamicScheduler, GreedyScheduler,
+                DynamicScheduler{OhMyThreads.Schedulers.NoChunking},
+                SerialScheduler, ChunkedGreedy)
                 @testset for split in (:batch, :scatter)
                     for nchunks in (1, 2, 6)
                         if sched == GreedyScheduler
                             scheduler = sched(; ntasks = nchunks)
                         elseif sched == DynamicScheduler{OhMyThreads.Schedulers.NoChunking}
-                            scheduler = DynamicScheduler(; chunking=false)
+                            scheduler = DynamicScheduler(; chunking = false)
                         elseif sched == SerialScheduler
                             scheduler = SerialScheduler()
                         else
@@ -30,7 +32,8 @@ ChunkedGreedy(;kwargs...) = GreedyScheduler(;kwargs...)
                         end
 
                         kwargs = (; scheduler)
-                        if (split == :scatter || sched ∈ (GreedyScheduler, ChunkedGreedy)) || op ∉ (vcat, *)
+                        if (split == :scatter ||
+                            sched ∈ (GreedyScheduler, ChunkedGreedy)) || op ∉ (vcat, *)
                             # scatter and greedy only works for commutative operators!
                         else
                             mapreduce_f_op_itr = mapreduce(f, op, itrs...)
@@ -66,7 +69,8 @@ end;
 @testset "ChunkSplitters.Chunk" begin
     x = rand(100)
     chnks = OhMyThreads.chunks(x; n = Threads.nthreads())
-    for scheduler in (DynamicScheduler(; chunking=false), StaticScheduler(; chunking=false))
+    for scheduler in (
+        DynamicScheduler(; chunking = false), StaticScheduler(; chunking = false))
         @testset "$scheduler" begin
             @test tmap(x -> sin.(x), chnks; scheduler) ≈ map(x -> sin.(x), chnks)
             @test tmapreduce(x -> sin.(x), vcat, chnks; scheduler) ≈
@@ -86,106 +90,106 @@ end;
 
     # reduction
     @test @tasks(for i in 1:3
-        @set reducer=(+)
+        @set reducer = (+)
         i
     end) == 6
 
     # scheduler settings
     for sched in (StaticScheduler(), DynamicScheduler(), GreedyScheduler())
         @test @tasks(for i in 1:3
-            @set scheduler=sched
+            @set scheduler = sched
             i
         end) |> isnothing
     end
     # scheduler settings as symbols
     @test @tasks(for i in 1:3
-        @set scheduler=:static
+        @set scheduler = :static
         i
     end) |> isnothing
     @test @tasks(for i in 1:3
-        @set scheduler=:dynamic
+        @set scheduler = :dynamic
         i
     end) |> isnothing
     @test @tasks(for i in 1:3
-        @set scheduler=:greedy
+        @set scheduler = :greedy
         i
     end) |> isnothing
 
     # @set begin ... end
     @test @tasks(for i in 1:10
         @set begin
-            scheduler=StaticScheduler()
-            reducer=(+)
+            scheduler = StaticScheduler()
+            reducer = (+)
         end
         i
     end) == 55
     # multiple @set
     @test @tasks(for i in 1:10
-        @set scheduler=StaticScheduler()
+        @set scheduler = StaticScheduler()
         i
-        @set reducer=(+)
+        @set reducer = (+)
     end) == 55
     # @set init
     @test @tasks(for i in 1:10
         @set begin
-            reducer=(+)
-            init=0.0
+            reducer = (+)
+            init = 0.0
         end
         i
     end) === 55.0
     @test @tasks(for i in 1:10
         @set begin
-            reducer=(+)
-            init=0.0*im
+            reducer = (+)
+            init = 0.0 * im
         end
         i
     end) === (55.0 + 0.0im)
 
     # top-level "kwargs"
     @test @tasks(for i in 1:3
-        @set scheduler=:static
-        @set ntasks=1
+        @set scheduler = :static
+        @set ntasks = 1
         i
     end) |> isnothing
     @test @tasks(for i in 1:3
-        @set scheduler=:static
-        @set nchunks=2
+        @set scheduler = :static
+        @set nchunks = 2
         i
     end) |> isnothing
     @test @tasks(for i in 1:3
-        @set scheduler=:dynamic
-        @set chunksize=2
+        @set scheduler = :dynamic
+        @set chunksize = 2
         i
     end) |> isnothing
     @test @tasks(for i in 1:3
-        @set scheduler=:dynamic
-        @set chunking=false
+        @set scheduler = :dynamic
+        @set chunking = false
         i
     end) |> isnothing
     @test_throws ArgumentError @tasks(for i in 1:3
-        @set scheduler=DynamicScheduler()
-        @set chunking=false
+        @set scheduler = DynamicScheduler()
+        @set chunking = false
         i
     end)
     @test_throws MethodError @tasks(for i in 1:3
-        @set scheduler=:serial
-        @set chunking=false
+        @set scheduler = :serial
+        @set chunking = false
         i
     end)
     @test_throws MethodError @tasks(for i in 1:3
-        @set scheduler=:dynamic
-        @set asd=123
+        @set scheduler = :dynamic
+        @set asd = 123
         i
     end)
 
     # TaskLocalValue
-    ntd = 2*Threads.nthreads()
+    ntd = 2 * Threads.nthreads()
     ptrs = Vector{Ptr{Nothing}}(undef, ntd)
     tids = Vector{UInt64}(undef, ntd)
     tid() = OhMyThreads.Tools.taskid()
     @test @tasks(for i in 1:ntd
         @local C::Vector{Float64} = rand(3)
-        @set scheduler=:static
+        @set scheduler = :static
         ptrs[i] = pointer_from_objref(C)
         tids[i] = tid()
     end) |> isnothing
@@ -217,17 +221,17 @@ end;
     sched = StaticScheduler()
     sched_sym = :static
     data = rand(10)
-    red = (a,b) -> a+b
+    red = (a, b) -> a + b
     n = 2
     @test @tasks(for d in data
-        @set scheduler=sched
-        @set reducer=red
+        @set scheduler = sched
+        @set reducer = red
         var * d
     end) ≈ var * sum(data)
     @test @tasks(for d in data
-        @set scheduler=sched_sym
-        @set ntasks=n
-        @set reducer=red
+        @set scheduler = sched_sym
+        @set ntasks = n
+        @set reducer = red
         var * d
     end) ≈ var * sum(data)
 
@@ -236,13 +240,14 @@ end;
     end
     @test @tasks(for _ in 1:10
         @local C = SingleInt(var)
-        @set reducer=+
+        @set reducer = +
         C.x
-    end) == 10*var
+    end) == 10 * var
 end;
 
 @testset "WithTaskLocals" begin
-    let x = TaskLocalValue{Base.RefValue{Int}}(() -> Ref{Int}(0)), y = TaskLocalValue{Base.RefValue{Int}}(() -> Ref{Int}(0))
+    let x = TaskLocalValue{Base.RefValue{Int}}(() -> Ref{Int}(0)),
+        y = TaskLocalValue{Base.RefValue{Int}}(() -> Ref{Int}(0))
         # Equivalent to
         # function f()
         #    x[][] += 1
@@ -259,7 +264,7 @@ end;
         # Make sure we can call `f` like a regular function
         @test f() == (1, 1)
         @test f() == (2, 2)
-        @test @fetch( f() ) == (1, 1)
+        @test @fetch(f()) == (1, 1)
         # Acceptable use of promise_task_local
         @test @fetch(promise_task_local(f)()) == (1, 1)
         # Acceptable use of promise_task_local
@@ -283,23 +288,32 @@ end;
 @testset "chunking mode + chunksize option" begin
     for sched in (DynamicScheduler, StaticScheduler)
         @test sched() isa sched
-        @test sched(; chunksize=2) isa sched
+        @test sched(; chunksize = 2) isa sched
 
-        @test OhMyThreads.Schedulers.chunking_mode(sched(; chunksize=2)) == OhMyThreads.Schedulers.FixedSize
-        @test OhMyThreads.Schedulers.chunking_mode(sched(; nchunks=2)) == OhMyThreads.Schedulers.FixedCount
-        @test OhMyThreads.Schedulers.chunking_mode(sched(; chunking=false)) == OhMyThreads.Schedulers.NoChunking
-        @test OhMyThreads.Schedulers.chunking_mode(sched(; nchunks=2, chunksize=4, chunking=false)) == OhMyThreads.Schedulers.NoChunking
-        @test OhMyThreads.Schedulers.chunking_mode(sched(; nchunks=-2, chunksize=-4, split=:whatever, chunking=false)) == OhMyThreads.Schedulers.NoChunking
-        @test OhMyThreads.Schedulers.chunking_enabled(sched(; chunksize=2)) == true
-        @test OhMyThreads.Schedulers.chunking_enabled(sched(; nchunks=2)) == true
-        @test OhMyThreads.Schedulers.chunking_enabled(sched(; nchunks=-2, chunksize=-4, chunking=false)) == false
-        @test OhMyThreads.Schedulers.chunking_enabled(sched(; nchunks=2, chunksize=4, chunking=false)) == false
+        @test OhMyThreads.Schedulers.chunking_mode(sched(; chunksize = 2)) ==
+              OhMyThreads.Schedulers.FixedSize
+        @test OhMyThreads.Schedulers.chunking_mode(sched(; nchunks = 2)) ==
+              OhMyThreads.Schedulers.FixedCount
+        @test OhMyThreads.Schedulers.chunking_mode(sched(; chunking = false)) ==
+              OhMyThreads.Schedulers.NoChunking
+        @test OhMyThreads.Schedulers.chunking_mode(sched(;
+            nchunks = 2, chunksize = 4, chunking = false)) ==
+              OhMyThreads.Schedulers.NoChunking
+        @test OhMyThreads.Schedulers.chunking_mode(sched(;
+            nchunks = -2, chunksize = -4, split = :whatever, chunking = false)) ==
+              OhMyThreads.Schedulers.NoChunking
+        @test OhMyThreads.Schedulers.chunking_enabled(sched(; chunksize = 2)) == true
+        @test OhMyThreads.Schedulers.chunking_enabled(sched(; nchunks = 2)) == true
+        @test OhMyThreads.Schedulers.chunking_enabled(sched(;
+            nchunks = -2, chunksize = -4, chunking = false)) == false
+        @test OhMyThreads.Schedulers.chunking_enabled(sched(;
+            nchunks = 2, chunksize = 4, chunking = false)) == false
 
-        @test_throws ArgumentError sched(; nchunks=2, chunksize=3)
-        @test_throws ArgumentError sched(; nchunks=0, chunksize=0)
-        @test_throws ArgumentError sched(; nchunks=-2, chunksize=-3)
+        @test_throws ArgumentError sched(; nchunks = 2, chunksize = 3)
+        @test_throws ArgumentError sched(; nchunks = 0, chunksize = 0)
+        @test_throws ArgumentError sched(; nchunks = -2, chunksize = -3)
 
-        let scheduler = sched(; chunksize=2)
+        let scheduler = sched(; chunksize = 2)
             @test tmapreduce(sin, +, 1:10; scheduler) ≈ mapreduce(sin, +, 1:10)
             @test tmap(sin, 1:10; scheduler) ≈ map(sin, 1:10)
             @test isnothing(tforeach(sin, 1:10; scheduler))
@@ -312,45 +326,151 @@ end;
     res_tmr = mapreduce(sin, +, 1:10000)
 
     # scheduler not given
-    @test tmapreduce(sin, +, 1:10000; ntasks=2) ≈ res_tmr
-    @test tmapreduce(sin, +, 1:10000; nchunks=2) ≈ res_tmr
-    @test tmapreduce(sin, +, 1:10000; split=:scatter) ≈ res_tmr
-    @test tmapreduce(sin, +, 1:10000; chunksize=2) ≈ res_tmr
-    @test tmapreduce(sin, +, 1:10000; chunking=false) ≈ res_tmr
+    @test tmapreduce(sin, +, 1:10000; ntasks = 2) ≈ res_tmr
+    @test tmapreduce(sin, +, 1:10000; nchunks = 2) ≈ res_tmr
+    @test tmapreduce(sin, +, 1:10000; split = :scatter) ≈ res_tmr
+    @test tmapreduce(sin, +, 1:10000; chunksize = 2) ≈ res_tmr
+    @test tmapreduce(sin, +, 1:10000; chunking = false) ≈ res_tmr
 
     # scheduler isa Scheduler
-    @test tmapreduce(sin, +, 1:10000; scheduler=StaticScheduler()) ≈ res_tmr
-    @test_throws ArgumentError tmapreduce(sin, +, 1:10000; ntasks=2, scheduler=DynamicScheduler())
-    @test_throws ArgumentError tmapreduce(sin, +, 1:10000; chunksize=2, scheduler=DynamicScheduler())
-    @test_throws ArgumentError tmapreduce(sin, +, 1:10000; split=:scatter, scheduler=StaticScheduler())
-    @test_throws ArgumentError tmapreduce(sin, +, 1:10000; ntasks=3, scheduler=SerialScheduler())
+    @test tmapreduce(sin, +, 1:10000; scheduler = StaticScheduler()) ≈ res_tmr
+    @test_throws ArgumentError tmapreduce(
+        sin, +, 1:10000; ntasks = 2, scheduler = DynamicScheduler())
+    @test_throws ArgumentError tmapreduce(
+        sin, +, 1:10000; chunksize = 2, scheduler = DynamicScheduler())
+    @test_throws ArgumentError tmapreduce(
+        sin, +, 1:10000; split = :scatter, scheduler = StaticScheduler())
+    @test_throws ArgumentError tmapreduce(
+        sin, +, 1:10000; ntasks = 3, scheduler = SerialScheduler())
 
     # scheduler isa Symbol
     for s in (:dynamic, :static, :serial, :greedy)
-        @test tmapreduce(sin, +, 1:10000; scheduler=s, init=0.0) ≈ res_tmr
+        @test tmapreduce(sin, +, 1:10000; scheduler = s, init = 0.0) ≈ res_tmr
     end
     for s in (:dynamic, :static, :greedy)
-        @test tmapreduce(sin, +, 1:10000; ntasks=2, scheduler=s, init=0.0) ≈ res_tmr
+        @test tmapreduce(sin, +, 1:10000; ntasks = 2, scheduler = s, init = 0.0) ≈ res_tmr
     end
     for s in (:dynamic, :static)
-        @test tmapreduce(sin, +, 1:10000; chunksize=2, scheduler=s) ≈ res_tmr
-        @test tmapreduce(sin, +, 1:10000; chunking=false, scheduler=s) ≈ res_tmr
-        @test tmapreduce(sin, +, 1:10000; nchunks=3, scheduler=s) ≈ res_tmr
-        @test tmapreduce(sin, +, 1:10000; ntasks=3, scheduler=s) ≈ res_tmr
-        @test_throws ArgumentError tmapreduce(sin, +, 1:10000; ntasks=3, nchunks=2, scheduler=s) ≈ res_tmr
+        @test tmapreduce(sin, +, 1:10000; chunksize = 2, scheduler = s) ≈ res_tmr
+        @test tmapreduce(sin, +, 1:10000; chunking = false, scheduler = s) ≈ res_tmr
+        @test tmapreduce(sin, +, 1:10000; nchunks = 3, scheduler = s) ≈ res_tmr
+        @test tmapreduce(sin, +, 1:10000; ntasks = 3, scheduler = s) ≈ res_tmr
+        @test_throws ArgumentError tmapreduce(
+            sin, +, 1:10000; ntasks = 3, nchunks = 2, scheduler = s)≈res_tmr
     end
 end;
 
 @testset "empty collections" begin
     for empty_coll in (11:9, Float64[])
-        for f in (sin, x->im*x, identity)
+        for f in (sin, x -> im * x, identity)
             for op in (+, *, min)
                 @test_throws ArgumentError tmapreduce(f, op, empty_coll)
-                for init in (0.0, 0, 0.0*im, 0f0)
+                for init in (0.0, 0, 0.0 * im, 0.0f0)
                     @test tmapreduce(f, op, empty_coll; init) == init
                 end
                 @test tforeach(f, empty_coll) |> isnothing
             end
+        end
+    end
+end;
+
+# for testing @one_by_one region
+mutable struct SingleAccessOnly
+    in_use::Bool
+    const lck::ReentrantLock
+    SingleAccessOnly() = new(false, ReentrantLock())
+end
+function acquire(f, o::SingleAccessOnly)
+    lock(o.lck) do
+        o.in_use && throw(ErrorException("Already in use!"))
+        o.in_use = true
+    end
+    try
+        f()
+    finally
+        lock(o.lck) do
+            !o.in_use && throw(ErrorException("Conflict!"))
+            o.in_use = false
+        end
+    end
+end
+
+@testset "regions" begin
+    @testset "@one_by_one" begin
+        sao = SingleAccessOnly()
+        try
+            @tasks for i in 1:10
+                @set ntasks = 10
+                @one_by_one begin
+                    acquire(sao) do
+                        sleep(0.01)
+                    end
+                end
+            end
+        catch ErrorException
+            @test false
+        else
+            @test true
+        end
+
+        # test escaping
+        x = 0
+        y = 0
+        sao = SingleAccessOnly()
+        try
+            @tasks for i in 1:10
+                @set ntasks = 10
+
+                y += 1 # not safe (race condition)
+                @one_by_one begin
+                    x += 1 # parallel-safe because inside of one_by_one region
+                    acquire(sao) do
+                        sleep(0.01)
+                    end
+                end
+            end
+            @test x == 10
+        catch ErrorException
+            @test false
+        end
+    end
+
+    @testset "@only_one" begin
+        x = 0
+        y = 0
+        try
+            @tasks for i in 1:10
+                @set ntasks = 10
+
+                y += 1 # not safe (race condition)
+                @only_one begin
+                    x += 1 # parallel-safe because only a single task will execute this
+                end
+            end
+            @test x == 1 # only a single task should have incremented x
+        catch ErrorException
+            @test false
+        end
+    end
+
+    @testset "@only_one + @one_by_one" begin
+        x = 0
+        y = 0
+        try
+            @tasks for i in 1:10
+                @set ntasks = 10
+
+                @only_one begin
+                    x += 1 # parallel-safe
+                end
+
+                @one_by_one begin
+                    y += 1 # parallel-safe
+                end
+            end
+            @test x == 1 &&  y == 10
+        catch ErrorException
+            @test false
         end
     end
 end;